Move icons and utility function. Move EntityTitle. Move project entity details. Add CollapsibleList components from EntityWithEverything
